server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Hyeonwoo KOO"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "google-api-python-client (>=2.181.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
